Rename group id to firebase.sdk.dashboard and add missing import statements to user resource.
