+++ src/main/java/firebase/sdk/dashboard/api/PlatformResource.java
@@ -1,4 +1,4 @@
-package com.google.capstone.api;
+package firebase.sdk.dashboard.api;
 
 import java.util.List;
 import java.util.HashMap;
@@ -10,9 +10,9 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.MediaType;
-import com.google.capstone.dao.PlatformReleaseDao;
-import com.google.capstone.dao.PlatformReleaseDaoDatastore;
-import com.google.capstone.dao.SDKDao;
+import firebase.sdk.dashboard.dao.PlatformReleaseDao;
+import firebase.sdk.dashboard.dao.PlatformReleaseDaoDatastore;
+import firebase.sdk.dashboard.dao.SDKDao;
 
 /**
  * Root resource (exposed at "myresource" path)
