+++ src/main/java/firebase/sdk/dashboard/api/UserResource.java
@@ -1,9 +1,11 @@
-package com.google.capstone.api;
+package firebase.sdk.dashboard.api;
 
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Arrays;
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.DELETE;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -11,9 +13,9 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
-import com.google.capstone.dao.PlatformReleaseDao;
-import com.google.capstone.dao.PlatformReleaseDaoDatastore;
-import com.google.capstone.dao.SDKDao;
+import firebase.sdk.dashboard.dao.PlatformReleaseDao;
+import firebase.sdk.dashboard.dao.PlatformReleaseDaoDatastore;
+import firebase.sdk.dashboard.dao.SDKDao;
 
 /**
  * Root resource (exposed at "myresource" path)
